- These are continued updates for #1193 - Moved the sink pixel flagging to occur AFTER the bias subtraction in wf3ccd. - Changed the flagging for sink pixels to represent what's in the text of the reference and not the example code. - Minor updates in cte to force more explicit types
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@41636 fe389314-cf27-0410-b35b-8c050e845b92

+++ pkg/wfc3/calwf3/wf3cte/cte.h
@@ -8,16 +8,16 @@ typedef struct {
     double cte_date0; /*date of uvis install on hst in mjd*/
     double cte_date1; /*date of cte model pinning mjd*/
     int cte_len; /*max length of cte trail */
-    float   rn_amp; /*read noise amplitude for clipping */
+    double   rn_amp; /*read noise amplitude for clipping */
     int n_forward; /* number of forward modeling iterations */
     int n_par; /*numver of iterations in parallel transfer */
     float scale_frac; /*scaling of cte model relative to ctedate1*/
     int noise_mit; /*read noise mitigation algorithm*/
-    float thresh; /*over subtraction threshold*/        
+    double thresh; /*over subtraction threshold*/        
     int cte_traps; /*number of valid TRAPS in file for reallocation*/
     int wcol_data[TRAPS]; /*trap number, insync with number of traps*/
-    float qlevq_data[TRAPS];/*charge packet size in electrons*/
-    float dpdew_data[TRAPS];/*trap size in electrons*/  
+    double qlevq_data[TRAPS];/*charge packet size in electrons*/
+    double dpdew_data[TRAPS];/*trap size in electrons*/  
     int   iz_data[RAZ_COLS]; /*column number in raz format*/
     double scale512[RAZ_COLS]; /*scaling appropriate at row 512 */
     double scale1024[RAZ_COLS];/*scaling appropriate at row 1024 */
@@ -72,10 +72,10 @@ int doCteBias (WF3Info *, SingleGroup *);
 int GetCTEFlags (WF3Info *, Hdr *);
 int a2d_raz(WF3Info *);
 int raw2raz(WF3Info *, SingleGroup *, SingleGroup *, SingleGroup *);
-int raz2rsz(WF3Info *, SingleGroup *, SingleGroup *, float , int );
+int raz2rsz(WF3Info *, SingleGroup *, SingleGroup *, double , int );
 int findPostScanBias(SingleGroup *, float *, float *);
 int findPreScanBias(SingleGroup *, float *, float *);
-int find_dadj(int ,int , float [][RAZ_ROWS], float [][RAZ_ROWS], float , float *);
+int find_dadj(int ,int , float [][RAZ_ROWS], float [][RAZ_ROWS], double , float *);
 int rsz2rsc(WF3Info *, SingleGroup *, SingleGroup *, CTEParams * );
 int inverse_cte_blur(SingleGroup *, SingleGroup *, SingleGroup *, CTEParams *, int, double);
 int sim_colreadout_l(double *, double *, double *, CTEParams *);
